+++ index.html
@@ -236,7 +236,6 @@
   </template>
 
 </body>
-<script src="js/functions.js"></script>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 
 </html>
